arch: the chunk codec leaves sync.py (#214)
The third slice of #201, and the one its own text said was a move rather than a
redesign: `pack`, `unpack`, `split_for_export` and the two size bounds import
`zlib` and nothing from the package, and no ordering argument touches them.
`woswoar/codec.py` is 134 lines; `sync.py` is 3063, from 3170.

Worth doing rather than filing because `unpack`'s bound is a claim in
`docs/security.md`, and claims there are backed by tests: until this moved,
reaching the codec from a test meant driving `run()`. `tests/test_codec.py` is
that collection now -- eleven tests that need no sandbox, no age and no git.
What stays in `tests/test_sync.py` is the half that needs a repository: a bomb
arriving in a real chunk, and the flushing that keeps a host's plaintext from
being held all at once.

A pure move, and checked rather than asserted:

    6 command(s) against main (07584aa), 67 lines
    scrubs beyond $HOME and $TREE:
      age1[a-z0-9]{58} -> AGE_KEY
      SHA256:[A-Za-z0-9+/]{43} -> FINGERPRINT
    identical

Two constraints the issue named, both real. `sync` reaches the codec by
attribute: eight tests patch `codec.MAX_EXPORT_BYTES` and `codec.pack` by
*string*, so binding the bare names would have unhooked every one of them --
`codec` joins `SPY_SEAMS` so that stays enforced rather than remembered. And
CI's attack job imports the name, so `.github/workflows/ci.yml` moves in the
same commit; a stale import there fails in the disguise of a security
regression. A caller the issue did not list turned up too: `prove` calls
`unpack` directly.

/simplify, applied: `docs/architecture.md` was the real finding. The canonical
map still described a package without `codec`, and its "where the shape is
under strain" section still argued for doing this work -- five places, none of
which CI can see, since `tests/test_docs.py` checks links rather than claims.
The module docstring also quoted a sentence that is not in `docs/security.md`;
it quotes the one that is. `sync.py`'s own "two layers did come out cleanly" is
three, and the seven comments there that still named `MAX_CHUNK_BYTES` and
`unpack` as though they were local now say where they went.
