+++ .github/workflows/ci.yml
@@ -329,7 +329,7 @@ jobs:
           git clone --quiet --no-local "$BASE/origin.git" "$ATTACK"
           DAY_PUB=$(cat "$ATTACK/hosts/$ALPHA_ID/keys/2023-11-14.pub")
           printf '1700000009\ts1\t~\t0\t5\tforged-marker-command\n' \
-            | python3 -c 'import sys; from woswoar import sync; sys.stdout.buffer.write(sync.pack(sys.stdin.buffer.read()))' \
+            | python3 -c 'import sys; from woswoar import codec; sys.stdout.buffer.write(codec.pack(sys.stdin.buffer.read()))' \
             | age -r "$DAY_PUB" \
             > "$ATTACK/hosts/$ALPHA_ID/2023-11-14/9999999999-ffffff.age"
           git -C "$ATTACK" add -A
